requirements.txt
paramiko~=3.4.0
requests~=2.31.0
PySocks~=1.7.1
